Add "realname" as a new connection parameter
You can use this to set the realname which is sent to the IRC server in
the initial USER command, and will be displayed in your WHOIS output.

Convos will append "via https://convos.chat" to the value you specify,
so as to have a simple and mostly-reliable way to detect other clients
using Convos.