fix segfaults in docker containers built for the wrong architecture
On certain architectures, when Crypt::Argon2 is built for
`-march=native` but then run on a different architecture (which is easy
to do when building on say github actions then deploying elsewhere) the
resulting build of the library will segfault when convos attempts to
login, killing the container.

Since Crypt::Argon2 version 0.22 an environment variable is available
that allows building without the `-march=native` optimization (when set
to NONE).